pyproject.toml
[project]
name = "desafio-tecnico-monks"
version = "0.1.0"
description = ""
authors = [
    {name = "Alekssandher",email = "94875660+Alekssandher@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "polars (>=1.33.1,<2.0.0)",
    "polars-lts-cpu (>=1.33.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "httpie (>=3.2.4,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)"
]
